Fixed test issues by moving routes to correct files
