+++ backend/routes/userRoutes.js
@@ -16,4 +16,11 @@ router.put("/:id", verifyToken, asyncHandler(userController.updateUser));
 
 router.delete("/:id", verifyToken, asyncHandler(userController.deleteUser));
 
+// Fetch tournaments a user is registered for
+router.get(
+  "/:id/tournaments",
+  verifyToken,
+  asyncHandler(userController.getUserTournaments)
+);
+
 module.exports = router;
